[Silver IV] Title: 피하자, Time: 524 ms, Memory: 146968 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/25379. 피하자/피하자.java
@@ -25,12 +25,12 @@ public class Main {
             }
         }
 
-        int oddToFront = 0;
+        long oddToFront = 0;
         for (int i = 0; i < odd.size(); i++) {
             oddToFront += Math.abs(odd.get(i) - i);
         }
 
-        int evenToFront = 0;
+        long evenToFront = 0;
         for (int i = 0; i < even.size(); i++) {
             evenToFront += Math.abs(even.get(i) - i);
         }
